array converters (#4)
Wrap the attrs conversion mechanism, don't use it, do conversions ourselves in __attrs_post_init__ after attrs init runs (so everything is in __dict__) before the data tree is initialized. This is the right time to convert e.g. sparse/list-based inputs to arrays before handing them to xarray.

We need to be careful not to pop anything out of __dict__ in converters. That is the responsibility of the tree init routine, which needs everything in there.