decorator inheritance wip
a type error was incorrectly raised for xattree classes having child fields whose type annotation is a base class, not the concrete child type. decorating both parent and child classes should now work properly. start work on inheritance of the decorator such that if a base class is decorated child classes don't have to be, as well as an option to turn this off.