double decoration fix (#19)
a type error was incorrectly raised for xattree classes having child fields whose type annotation is a base class, not the concrete child type, when the base class is also xattree. this should now work.

also start work on inheritance of the decorator such that if a base class is decorated child classes don't have to be, as well as an option to turn this off.